Update ssa.py
Added the self.till_end attribute for the user to choose if he wants the simulation to stop when no valid reactions are found or if he wants to continue till the end of time or time steps he indicated. In case he chooses to don't stop the simulation but doesn't indicate desired end time or total time steps, raises an exception.